xrandr: fix crash if some modes cannot be found
When printing modes in "xrandr -q", check to see if we failed to look up
mode information from a mode XID.  Previously the command would
dereference null and crash if the mode information could not be found.

When using an external HDMI monitor on a laptop with a Skylake Intel
graphics chipset "xrandr -q" occasionally is unable to look up mode
information for some of the modes.  It seems likely there is some other
sort of library or driver issue causing these lookup failures, but this
change to xrandr at least prevents it from segfaulting.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>